Move recipes methods into Fatsecret class
